Move application test
